requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
PyPDF2==3.0.1
Pillow==10.2.0
anyio==4.3.0 
packaging==23.2 
tenacity==8.2.3 
docutils==0.20.1
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.3.0
pypdf
pymupdf
langchain-google-genai>=0.3.0
google-generativeai
faiss-cpu
